requirements.txt
Flask==1.1.1
flask-restful==0.3.7
flask_cors
Appium-Python-Client==0.49
selenium
pytest
